changed all folder and their respective locations in loginsignup branch
