+++ CSS/login.css
@@ -49,8 +49,8 @@
     margin-top: 15px;
 }
 
-input[type="radio"]{
-    cursor:pointer;
+input[type="radio"] {
+    cursor: pointer;
 }
 
 .sel {
