+++ Components/emailpreference.js
@@ -7,7 +7,7 @@ let emailpref = () => {
             <a class="anchor" href="./orderhistory.html">Order History</a><br>
             <a class="anchor" href="./appointments.html">Appointments</a><br>
             <a class="anchor" href="./addresses.html">Addresses</a><br>
-            <a class="anchor" href="/Unit-3-Project/Login/pages/login.html">Logout</a>
+            <a class="anchor" href="./login.html">Logout</a>
         </div>
         <div id="acchamber">
             <h4 class="acdetail">Email Preferences</h4><br>
