+++ HTML/accounts.html
@@ -1,16 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../styles/addresses.css">
-    <title>Addresses | SSENSE</title>
+    <link rel="stylesheet" href="../CSS/accounts.css">
+    <title>Account | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
 </head>
+
 <body>
-    
+    <div id="scontainer">
+        
+    </div>
 </body>
+
 </html>
-<script src="../scripts/addresses.js" type="module"></script>
+<script src="../JS/accounts.js" type="module"></script>
