+++ HTML/addresses.html
@@ -1,21 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Email Preference | SSENSE</title>
-    <link rel="stylesheet" href="../styles/emailpreference.css">
-    <link rel="stylesheet" href="../styles/accounts.css">
+    <link rel="stylesheet" href="../CSS/addresses.css">
+    <title>Addresses | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
 </head>
-
 <body>
-    <div id="scontainer">
-    </div>
+    
 </body>
-
 </html>
-<script src="../scripts/emailpreference.js" type="module"></script>
+<script src="../JS/addresses.js" type="module"></script>
