+++ HTML/appointments.html
@@ -4,15 +4,13 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Order History | SSENSE</title>
-    <link rel="stylesheet" href="../styles/accounts.css">
-    <link rel="stylesheet" href="../styles/orderhistory.css">
+    <link rel="stylesheet" href="../CSS/appointments.css">
+    <title>Appointments | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
 </head>
 <body>
-    <div id="scontainer">
-    </div>
+    
 </body>
 </html>
-<script src="../scripts/orderhistory.js" type="module"></script>
+<script src="../JS/appointments.js" type="module"></script>
