+++ HTML/emailpreference.html
@@ -1,21 +1,21 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../styles/login.css">
-    <title>CreateAccount | SSENSE</title>
+    <title>Email Preference | SSENSE</title>
+    <link rel="stylesheet" href="../CSS/emailpreference.css">
+    <link rel="stylesheet" href="../CSS/accounts.css">
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
 </head>
 
 <body>
-    <div id="loginchamber">
+    <div id="scontainer">
     </div>
 </body>
 
 </html>
-<script src="../scripts/passlogin.js" type="module"></script>
+<script src="../JS/emailpreference.js" type="module"></script>
