+++ HTML/landingpage.html
@@ -1,16 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../styles/appointments.css">
-    <title>Appointments | SSENSE</title>
+    <title>Landing Page | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
+    <link rel="stylesheet"
+        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
 </head>
+
 <body>
-    
+    <h1 style="font-family:'Poppins', sans-serif">Landing Page</h1>
 </body>
+
 </html>
-<script src="../scripts/appointments.js" type="module"></script>
