+++ HTML/login.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../styles/login.css">
+    <link rel="stylesheet" href="../CSS/login.css">
     <title>Login | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
@@ -19,4 +19,4 @@
 </body>
 
 </html>
-<script src="../scripts/login.js" type="module"></script>
+<script src="../JS/login.js" type="module"></script>
