+++ HTML/orderhistory.html
@@ -4,13 +4,15 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Landing Page | SSENSE</title>
+    <title>Order History | SSENSE</title>
+    <link rel="stylesheet" href="../CSS/accounts.css">
+    <link rel="stylesheet" href="../CSS/orderhistory.css">
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
 </head>
 <body>
-    <h1 style="font-family:'Poppins', sans-serif">Landing Page</h1>
+    <div id="scontainer">
+    </div>
 </body>
 </html>
+<script src="../JS/orderhistory.js" type="module"></script>
