+++ HTML/passlogin.html
@@ -1,11 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../styles/login.css">
-    <title>SignUp | SSENSE</title>
+    <link rel="stylesheet" href="../CSS/login.css">
+    <title>Login Account | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
     <link rel="stylesheet"
@@ -18,5 +19,4 @@
 </body>
 
 </html>
-
-<script src="../scripts/signup.js" type="module"></script>
+<script src="../JS/passlogin.js" type="module"></script>
