+++ HTML/signup.html
@@ -5,17 +5,19 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../styles/accounts.css">
-    <title>Account | SSENSE</title>
+    <link rel="stylesheet" href="../CSS/login.css">
+    <title>SignUp | SSENSE</title>
     <link rel="shortcut icon" href="https://res.cloudinary.com/ssenseweb/image/upload/v1472005257/web/favicon.ico"
         type="image/x-icon">
+    <link rel="stylesheet"
+        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
 </head>
 
 <body>
-    <div id="scontainer">
-        
+    <div id="loginchamber">
     </div>
 </body>
 
 </html>
-<script src="../scripts/accounts.js" type="module"></script>
+
+<script src="../JS/signup.js" type="module"></script>
