+++ JS/accounts.js
@@ -1,4 +1,4 @@
-import {account} from "../components/accounts.js";
+import {account} from "../Components/accounts.js";
 
 let container = document.getElementById("scontainer");
 container.innerHTML = account();
