+++ JS/emailpreference.js
@@ -1,4 +1,4 @@
-import { emailpref } from "../components/emailpreference.js";
+import { emailpref } from "../Components/emailpreference.js";
 
 let container = document.getElementById("scontainer");
 container.innerHTML = emailpref();
