+++ JS/orderhistory.js
@@ -1,4 +1,4 @@
-import { orderhistory } from "../components/orderhistory.js";
+import { orderhistory } from "../Components/orderhistory.js";
 
 let container = document.getElementById("scontainer");
 container.innerHTML = orderhistory();
